chore: refactor embedder for improved readability
