+++ packages/server/tsconfig.json
@@ -8,15 +8,8 @@
       "parabol-client/*": ["client/*"],
       "~/*": ["client/*"]
     },
-    "lib": [
-      "esnext",
-      "dom"
-    ],
-    "types": [
-      "node",
-      "jest",
-      "jest-extended"
-    ]
+    "lib": ["esnext", "dom"],
+    "types": ["node", "jest", "jest-extended"]
   },
 
   "exclude": [
@@ -29,11 +22,11 @@
     "types/webpackEnv.ts",
     "types/modules.d.ts",
     "server.ts",
-    "../client/modules/email/components/SummaryEmail/MeetingSummaryEmail/MeetingSummaryEmail.tsx"
+    "../client/modules/email/components/SummaryEmail/MeetingSummaryEmail/MeetingSummaryEmail.tsx",
+    "../embedder/orgIdsWithFeatureFlag.ts",
+    "../embedder/unnestedArray.ts"
   ],
-  "include": [
-    "graphql/**/*.ts",
-  ]
+  "include": ["graphql/**/*.ts"]
 
   // if "include" or "files" is added, even if they are empty arrays, then strictNullChecks breaks
   // repro: https://www.typescriptlang.org/play?strictFunctionTypes=false&strictPropertyInitialization=false&strictBindCallApply=false&noImplicitThis=false&noImplicitReturns=false&alwaysStrict=false&declaration=false&experimentalDecorators=false&emitDecoratorMetadata=false&target=6&ts=3.5.1#code/C4TwDgpgBA8gRgKygXigbwFBSgWwIYhwQDKwATgIJlkD8AXFAM7kCWAdgOYDaAuhgL4ZQkKFTIpYiLgHJ8hEuTHS+AYwD2bZlDwS2AVwA2B7Y21sQJ0dQx4AdADM1ZAKJ4VACwAUngF4BKFAA+KH8MIA
